+++ src/test/php/lang/ast/unittest/emit/FunctionTypesTest.class.php
@@ -7,6 +7,7 @@ use lang\Primitive;
  * Function types
  *
  * @see  https://docs.hhvm.com/hack/types/summary-table
+ * @see  https://docs.hhvm.com/hack/callables/introduction
  */
 class FunctionTypesTest extends EmittingTest {